requirements.txt
pyautogui
pillow
opencv-python
numpy
mss
scikit-learn
pynput
keyboard